% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown_age.R
\name{crown_age}
\alias{crown_age}
\title{Calculate the crown age of a phylogeny}
\usage{
crown_age(phylo)
}
\arguments{
\item{phylo}{Object of class \code{phylo} or \code{multiPhylo}. A rooted ultrametric phylogeny of class \code{phylo} or a set of rooted ultrametric phylogenies of class \code{multiPhylo}. The rooted ultrametric phylogenie(s) can have polytomie(s) (i.e. non binary tree).}
}
\value{
Returns the value of the crown age in the same unit as the branch length of the phylogenie(s) (typically in million of years).
}
\description{
Calculate the crown age of a rooted ultrametric phylogeny or set of phylogenies. This function can used on stem or crown phylogenie(s) and on binary or non binary phylogenie(s).
}
\details{
This function will calculate the crown age of a phylogeny or a set of phylogenies.
}
\seealso{
\code{\link{stem_age}}
}
\author{
Sophia Lambert
}
