% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_Rho.R
\name{int_Rho}
\alias{int_Rho}
\title{Integrate function}
\usage{
int_Rho(dt)
}
\arguments{
\item{dt}{Numeric. The length of the intervals on which the function is assumed to be constant. If \code{dt = 0} the function is integrated using  \code{stats::integrate} while overpassing errors. If a positive \code{dt} is given as argument, the integral is computed using a piece-wise constant approximation.}

\item{f}{Function. An \code{R} function of one variable on which to integrate.}

\item{lower}{Numeric. Lower limit of integration.}

\item{upper}{Numeric. Upper limit of integration.}

\item{rel.tol}{Numeric. Relative accuracy requested.}
}
\value{
The final estimate of the integral.
}
\description{
Integrate function either from \code{stats::integrate} while overpassing errors or manual integration.
}
\details{
If \code{dt = 0} any other arguments of \code{stats::integrate} can be specified. This function is a closure, it takes \code{dt} as argument and creates another function with arguments \code{f}, \code{lower}, \code{upper} and \code{rel.tol}.
}
\author{
Sophia Lambert
}
