% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P1_a.R
\name{P1_a}
\alias{P1_a}
\title{Probabillity that all speciation and extinction events happened before a certain time.}
\usage{
P1_a(t, r, epsi)
}
\arguments{
\item{t}{Numeric The specific time in the past before which all the events happened in millions of years.}

\item{r}{Numeric. The net diversification rate: \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}

\item{epsi}{Numeric. The turnover rate: \eqn{\mu / \lambda}. The units are in events/lineages/millions years.}
}
\value{
The probability that all birth and death events of the birth-death process happened before a time \eqn{t}.
}
\description{
Calculating the probabillity that all speciation and extinction events happened before a certain time \eqn{t}.
}
\seealso{
\code{\link{likelihood_bdK}}, \code{\link{fitMCMC_bdK}}
}
\author{
Sophia Lambert
}
