% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pa.R
\name{Pa}
\alias{Pa}
\title{Probabillity that all speciation and extinction events happened after a certain time.}
\usage{
Pa(t, r, epsi)
}
\arguments{
\item{t}{Numeric The specific time in the past after which all the events happened in millions of years. Typically, the age of the clade is used to condition on the start of the process.}

\item{r}{Numeric. The net diversification rate: \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}

\item{epsi}{Numeric. The turnover rate: \eqn{\mu / \lambda}. The units are in events/lineages/millions years.}
}
\value{
The probability that all birth and death events of the birth-death process happened after a time \eqn{t}.
}
\description{
Calculating the probabillity that all speciation and extinction events happened after a certain time \eqn{t}.
}
\seealso{
\code{\link{likelihood_bdK}}, \code{\link{fitMCMC_bdK}}
}
\author{
Sophia Lambert
}
