% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcond.R
\name{Pcond}
\alias{Pcond}
\title{Conditioning the birth-death process on a certain time and the survival of the process.}
\usage{
Pcond(t, r, epsi, y)
}
\arguments{
\item{t}{Numeric. The specific time in the past in millions of years. Typically, the age of the clade is used to condition on the start of the process.}

\item{r}{Numeric. The net diversification rate: \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}

\item{epsi}{Numeric. The turnover rate: \eqn{\mu / \lambda}. The units are in events/lineages/millions years.}

\item{y}{Numeric. The sampling probability also called sampling fraction: \eqn{k / N}.}
}
\value{
The probability of the birth-death process starting with one lineage at time \eqn{t} and surviving at present.
}
\description{
Calculating the probability of conditioning the birth-death process on the time \eqn{t} and its survival to present day.
}
\seealso{
\code{\link{likelihood_bdRho}}, \code{\link{fitMCMC_bdRho}}
}
\author{
Sophia Lambert
}
