% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcond_reparam.R
\name{Pcond_reparam}
\alias{Pcond_reparam}
\title{Conditioning the birth-death process on a certain time and the survival of the process reparametrised.}
\usage{
Pcond_reparam(t, yl, r)
}
\arguments{
\item{t}{Numeric. The specific time in the past in millions of years. Typically, the age of the clade is used to condition on the start of the process.}

\item{yl}{Numeric. The sampling probability times the speciation rate: \eqn{y*\lambda}.}

\item{r}{Numeric. The net diversification rate: \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}
}
\value{
The probability of the birth-death process starting with one lineage at time \eqn{t} and surviving at present parametrised on \eqn{y*\lambda} and \eqn{r}.
}
\description{
Calculating the probability of conditioning the birth-death process  on the time \eqn{t} and its survival to present day parametrised on \eqn{y*\lambda} and \eqn{r}.
}
\seealso{
\code{\link{likelihood_bdRho}}, \code{\link{fitMCMC_bdRho}}
}
\author{
Sophia Lambert
}
