% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pnd.R
\name{Pnd}
\alias{Pnd}
\title{Probability density of node depths.}
\usage{
Pnd(t, r, epsi, y)
}
\arguments{
\item{t}{Numeric vector. A vector specifying the node depths of a phylogeny in millions of years.}

\item{r}{Numeric. The net diversification rate: \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}

\item{epsi}{Numeric. The turnover rate: \eqn{\mu / \lambda}. The units are in events/lineages/millions years.}

\item{y}{Numeric. The sampling probability also called sampling fraction: \eqn{k / N}.}
}
\value{
The probability density of node depths equivalent to \eqn{f\_y(t)} in Lambert 2018 and to \eqn{\lambda * p1} in Stadler 2010 and Lambert et al. 2022.
}
\description{
Calculating the probability density of node depths of a phylogenetic tree.
}
\references{
Lambert, A. (2018). The coalescent of a sample from a binary branching process. Theoretical population biology, 122, 30-35.
Stadler, T. (2010). Sampling-through-time in birth–death trees. Journal of theoretical biology, 267(3), 396-404.
}
\seealso{
\code{\link{likelihood_bdRho}}, \code{\link{fitMCMC_bdRho}}
}
\author{
Sophia Lambert
}
