% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcond.R
\name{Pcond}
\alias{Pcond}
\title{Conditioning of the birth-death process on the age and the survival of the process.}
\usage{
Pcond(t, r, epsi, y)
}
\arguments{
\item{t}{Numeric. The time of the start of the birth-death process in millions of years.}

\item{r}{Numeric. The net diversification rate : \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}

\item{epsi}{Numeric. The turnover rate : \eqn{\mu / \lambda}. The units are in events/lineages/millions years.}

\item{y}{Numeric. The sampling probability also called sampling fraction : \eqn{k / N}.}
}
\value{
The probability of the birth-death process starting with one lineage at time t and surviving at present.
}
\description{
Calculating the conditioning of the birth-death process on the age of the process and its survival.
}
\author{
Sophia Lambert
}
