% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcond_reparam.R
\name{Pcond_reparam}
\alias{Pcond_reparam}
\title{Conditioning the birth-death process on the age and the survival of the process reparametrised.}
\usage{
Pcond_reparam(t, yl, r)
}
\arguments{
\item{t}{Numeric. The time of the start of the birth-death process in millions of years.}

\item{yl}{Numeric. The sampling probability times the speciation rate : \eqn{y*\lambda}.}

\item{r}{Numeric. The net diversification rate : \eqn{\lambda - \mu}. The units are in events/lineages/millions years.}
}
\value{
The probability of the birth-death process starting with one lineage at time t and surviving at present parametrised on \eqn{y*\lambda} and \eqn{r}.
}
\description{
Calculating the probability of conditioning the birth-death process on the age of the process and its survival parametrised on \eqn{y*\lambda} and \eqn{r}.
}
\author{
Sophia Lambert
}
