% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi}
\alias{phi}
\title{Prior distribution on sampling probabilitie(s)}
\usage{
phi(beta = FALSE, unif = FALSE)
}
\arguments{
\item{beta}{Logical. If TRUE, a beta distribution is assumed on the sampling probabilitie(s) \eqn{y}.}

\item{unif}{Logical. If TRUE, a uniform distribution from a to b is assumed on the sampling probabilitie(s) \eqn{y}.}
}
\value{
Returns an object of class \code{"function"}. This function takes the following arguments and will return the probability density of the sampling probabilitie(s) \eqn{y} assuming a beta distribution or a uniform distribution from a to b:
\describe{
  \item{x}{Numeric vector. The value of the sampling probabilitie(s) \eqn{y}.}
  \item{a}{Numeric. The value of \eqn{\alpha} or \eqn{a} respectively for the beta or the uniform distribution. This value cannot be negative for both distributions and cannnot exceed \eqn{b} for the uniform distribution.}
  \item{b}{Numeric. The value of \eqn{\beta} or \eqn{b} respectively for the beta or the uniform distribution. This value cannot be negative for both distributions, cannnot be inferior to \eqn{a} and cannot exceed 1 for the uniform distribution.}
}
}
\description{
Calculating the probability density function of the sampling probabilitie(s) \eqn{y} assuming a beta distribution or a uniform distribution from a to b.
}
\details{
This function is a closure. This functions takes \code{beta} and \code{unif} as arguments and creates another function with arguments \code{x}, \code{a} and \code{b}.
}
\seealso{
\code{\link{likelihood_bdRho}}, \code{\link{fitMCMC_bdRho}}
}
\author{
Sophia Lambert
}
